better generation logic
